remove implicit conversions

+++ lib/murmurhash.h
@@ -47,7 +47,7 @@ static inline uint32_t murmur_32_scramble(uint32_t k)
  * @param seed A seed, for whom matched with a given key, makes the hash function produce the same hash for the key.
  * @return A `uint32_t`-sized hash of the bytes.
  */
-static inline uint32_t murmur3_32(const uint8_t* key_ptr, const size_t len, const uint32_t seed)
+static inline uint32_t murmur3_32(const uint8_t* key_ptr, const uint32_t len, const uint32_t seed)
 {
     uint32_t h = seed;
     uint32_t k;
